Update main.go
Updated the main.go code to conditionally use the provided kubeconfig file instead of trying to create an in-cluster configuration. The `createK8sClient` function creates a Kubernetes client using the provided kubeconfig file. This client is used to interact with the Kubernetes API server to fetch node information and create resources.